Better fix for pressing enter on a form
if we bind to the form inputs then it's possible to press enter in a situation where focus is inside the form but not an a field. Better to catch keyboard events for the whole form and then check the target and allow enter presses only for the inputs we want to handle them.